Add SIMD subgroup charter
As this is the first time we have separate proposals under the same subgroup - not sure where the charter should live. For now, I'm duplicating the charter here as well as in the relaxed-simd proposal for discoverability, and easy modification in case we decide to split the subgroup based on proposals in the future.